requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
redis==5.0.7
requests==2.32.3
smmap==5.0.1
spotipy==2.24.0
urllib3==2.2.2
